test slice pushdown optimization
